+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Minor] Reworked CPX-E class structure.
 
 ### Fixed
+- [Patch] Bugfixes for IO-Link modules
 - [Patch] Fixed diagnostics function in cpx-e-8do
 - [Patch] Fixed several linter issues
 - [Patch] Bugfixes after Merge